requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.3
openpyxl==3.1.2
psycopg2-binary==2.9.9

